feat: EdgeBuffer dtt_shortest_edge
Returns component to k edges priority queue. HDBSCAN with this dual tree traversal method is still slower than its single tree counterpart. However, the dual tree one seems to scale much better as the components size grows and n_components decreases at each iteration of the algorithm, which is promising. The next step will be to find where the meat of the overhead is in the EdgesBuffer cache.